feature(compass-saved-aggregations-queries): clean up
